requirements.txt
fastapi==0.115.14
elasticsearch==9.0.2
sentence-transformers==4.1.0
uvicorn[standard]==0.32.1
google-generativeai==0.8.3
python-dotenv==1.0.0